requirements.txt
streamlit
torch
torchvision
opencv-python-headless
numpy
pdf2image
pytesseract
Pillow
pandas